вынес дополнительные функции в helpers.js
