+++ src/components/AddStudent/Connector.js
@@ -1,0 +1,12 @@
+import { normalizeText } from 'utils/helpers';
+import Controller from './Controller';
+
+function Connector() {
+    return (
+       <Controller 
+            normalizeText={normalizeText}
+       />
+    );
+}
+
+export default Connector;
