+++ src/components/AddTeacher/Connector.js
@@ -1,0 +1,13 @@
+import { normalizePhone, normalizeText } from 'utils/helpers';
+import Controller from './Controller';
+
+function Connector() {
+    return (
+       <Controller 
+            normalizePhone={normalizePhone}
+            normalizeText={normalizeText}
+       />
+    );
+}
+
+export default Connector;
