+++ src/components/AddTeacher/Controller.js
@@ -2,7 +2,7 @@ import { useRef } from 'react';
 import sendRequest from 'api/utils';
 import View from './View';
 
-function Controller() {
+function Controller({normalizeText, normalizePhone}) {
     const nameRef = useRef(null);
     const lastnameRef = useRef(null);
     const patronymicRef = useRef(null);
@@ -22,25 +22,6 @@ function Controller() {
         teacherRef,
         adminRef
     };
-
-    const normalizeText = (value) => {
-        if (!value) return '';
-        const val = value.replace(/[^а-яёА-ЯЁ]/u, '');
-        if (val.length > 0) return val[0].toUpperCase() + val.slice(1).toLowerCase();
-        return '';
-    }
-
-    const normalizePhone = (value) => {
-        if (!value) return '';
-        let val;
-        if (value.length === 1) val = value.replace(/[^\d]/g, '');
-        else val = value.replace(/[^\d]/g, '').slice(1);
-        if (val.length < 4) {
-            return `+7 (${val}`;
-        }
-        if (val.length < 7) return `+7 (${val.slice(0, 3)}) ${val.slice(3)}`;
-        return `+7 (${val.slice(0, 3)}) ${val.slice(3, 6)}-${val.slice(6, 10)}`;
-    }
 
     function createTeacher(){
         const data = {
