+++ src/components/AddTeacher/index.js
@@ -1,2 +1,2 @@
 // eslint-disable-next-line no-restricted-exports
-export {default} from './Controller';
+export {default} from './Connector';
