+++ src/pages/InstitutionAdmin/Connector.js
@@ -1,5 +1,4 @@
 import { useSelector } from 'react-redux';
-
 import { useState } from 'react';
 
 import Controller from './Controller';
